(PC-19704) feat(homepage): create a adapted homepage fixture
